Fix handling of constants: constants were requiring coersion wrappers. Changed .v test cases so that they would not result in boxing (to be like Chapel).
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1478 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ test/flowanalysis/jplevyak/classvar.v
@@ -1,7 +1,7 @@
 
 class Vars {
-  var x : int;
-  var y : float;
+  var x : float;
+  var y : int;
 };
 
 v : new Vars;
